requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy>=1.26.0,<2.0.0
scipy>=1.11.0
scikit-image>=0.22.0
Pillow>=10.1.0
python-dotenv==1.0.0